pyproject.toml
[project]
name = "skincancer-cnn-from-scratch"
version = "0.1.0"
description = ""
authors = [
    {name = "HowardKuo2025",email = "howard.futurenest@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "visualkeras (>=0.2.0,<0.3.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "pydot (>=4.0.1,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
